Use i32 for index type in flang runtime

+++ flang/include/flang/ISO_Fortran_binding.h
@@ -40,7 +40,7 @@ typedef unsigned char CFI_rank_t;
 /* This type is probably larger than a default Fortran INTEGER
  * and should be used for all array indexing and loop bound calculations.
  */
-typedef ptrdiff_t CFI_index_t;
+typedef long CFI_index_t;
 
 typedef unsigned char CFI_attribute_t;
 #define CFI_attribute_pointer 1
